Refaktorert: ny komponent for vakter. Mulig å redigere ekstra rad.
